Waste less space on editor sample list index in GMDLS_val

+++ 4klang_source/Go4kVSTi/source/GoSynth/Go4kVSTiCore.h
@@ -293,7 +293,7 @@ typedef struct GMDLS_val
 	DWORD fileOffset;
 	BYTE flags;
 // GUI STUFF
-	DWORD sampleEntryListIndex;
+	WORD sampleEntryListIndex;
 } *GMDLS_valP;
 
 /////////////////////////////////////////////////////////////////////////////////////
